refactor(regular_expression): correct typo (#5429)
Just correct a misspelling.
