+++ crates/oxc_regular_expression/src/surrogate_pair.rs
